requirements.txt
streamlit
datetime
pandas
plotly.express
openpyxl
streamlit_extras